refactor: move ReviewHistorySkeleton (issue 17)
